Move a test into serialization module so it's easier to maintain
