+++ tests/histogram.rs
@@ -6,7 +6,6 @@ extern crate rand;
 use self::rand::Rng;
 
 use hdrsample::{Counter, Histogram, SubtractionError};
-use hdrsample::serialization::Serializer;
 use std::borrow::Borrow;
 use std::fmt;
 
@@ -534,27 +533,6 @@ fn total_count_overflow_from_add_with_resize_saturates() {
     h.add(h2).unwrap();
     assert_eq!(u64::max_value(), h.count_at(1));
     assert_eq!(u64::max_value(), h.len());
-}
-
-#[test]
-#[cfg(feature = "serialization")]
-fn total_count_overflow_from_deserialize_saturates() {
-    use hdrsample::serialization::{Deserializer, V2Serializer};
-    let mut h = Histogram::<u64>::new_with_bounds(1, u64::max_value(), 3).unwrap();
-
-    // can't go bigger than i64 max because it will be serialized
-    h.record_n(1, i64::max_value() as u64).unwrap();
-    h.record_n(1000, i64::max_value() as u64).unwrap();
-    h.record_n(1_000_000, i64::max_value() as u64).unwrap();
-    assert_eq!(u64::max_value(), h.len());
-
-    let mut vec = Vec::new();
-
-    V2Serializer::new().serialize(&h, &mut vec).unwrap();
-    let deser_h: Histogram<u64> = Deserializer::new()
-        .deserialize(&mut vec.as_slice())
-        .unwrap();
-    assert_eq!(u64::max_value(), deser_h.len());
 }
 
 #[test]
